Update patch set 2
Patch Set 2: -Code-Review

Ok, so I've updated the patch so that a Run launch will always use GdbProcessFactory.  This means a new Run launch and a debug launch created for CDI and then used for Run.  I think this is safe as the only real change in using this factory is the fact that we print the exit code.

Ultimately, probably for CDT 9.0, we'll need to better consolidate the Run part of LocalCDILaunchDelegate with DSF-GDB.

Patch-set: 2