Update patch set 6
Patch Set 6:

(2 comments)

This looks like a very nice solution that allows to extend things very easily!

I'm playing with it to see if we can simplify even further.

One thing that is risky is the creation of the three adapters that are used by other adapters through the get* methods (getSteppingController(), getSteppingModeTarget() and getViewModelAdapter()).  What bothers me is that these three _must_ be created before any other adapter that uses them.  The only thing that decides that is the order the adatpers are listed in getModelAdapters().  If a user were to override that method to add a new adapter at the beginning, and that the new adapter called getSteppingController(), the stepping controller would not be available yet.

We could put a comment in getModelAdapters() to specify the order, and assume that extenders would notice such a bug quickly.

We could also use getSteppingController() to actually create the controller if it was still null when this method was called.

Patch-set: 6
Label: Verified=0