Update patch set 8
Patch Set 8:

Pathset 8 aligns the use of generics with the new signatures of IAdapterFactory.

In GdbSessionAdapters, to use generics instead of Object, we need to cast to generic type T which could cause a ClassCastException.  We also must use
  @SuppressWarnings("unchecked")

I believe this is still better than using Object because if we were to return the wrong adapter, the ClassCastException will indicate this right away, instead of higher up the chain, when the adapter is used with a wrong type.

What do you think?

Patch-set: 8
Label: Verified=0