Update patch set 12
Patch Set 12:

I noticed that with the changes, we could no longer run the JUnit tests Headless.  The reason is that GdbLaunch.terminate() at the end of each test, which had the line:
  Platform.getAdapterManager().loadAdapter(this, adapter.getName());

This triggered the loading of the UI plugins, which then threw exceptions when running headless.

This was not a problem before because GdbLaunch.terminate() has special code to handle getting the ITerminateHanlder adapter, and therefore never got to the line:
Platform.getAdapterManager().loadAdapter(this, adapter.getName());

I looked at the history of this, all the way back to when it was first checked in to the DD project, and that loadAdapter() line has always been there, but I don't know why.

Patchset 12 removes it.  I'm not sure if that could cause problems so I left that line as a comment, so we'd know where to look if something failed down the line.

Mikhail, do you have any opinion on this?

Patch-set: 12
Label: Verified=0