Update patch set 2
Patch Set 2:

Patchset 2 does better handling of the RUN_MODE case.  The cleanup seems to also fix the fact that the launch was staying collapsed in the Debug view.

However, I'm wondering if extending GdbLaunchDelegate to support Run is confusing as it is named 'Gdb...'.  Maybe the better approach would be to create a delegate just for the Run mode.  I'll try it to see how it would look.

Patch-set: 2