Update patch set 2
Patch Set 2:

It is not ILaunch. In our case it is actually a String with path.
There are multiple participants, it should not throw an exception
if it does not know how to deal with object because it is propagated to the platform. Why do you want an exception?

Patch-set: 2
Reviewer: Gerrit User 592 <592@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0