Update patch set 2
Patch Set 2: Code-Review+1

> It is not ILaunch. In our case it is actually a String with path.
 > There are multiple participants, it should not throw an exception
 > if it does not know how to deal with object because it is
 > propagated to the platform. Why do you want an exception?

I don't know if the exception causes a special behaviour that we would lose.

As an example, maybe returning null causes no editor to open, while throwing an exception causes the special editor but the buttons to find the source code to open.  I just don't know this area well enough to know.

If you feel comfortable in removing the exception, then I'm fine with it.

Patch-set: 2