Speed up the execution of the Debug tests.
I noticed that there is some time spend terminating the launch.  That
time delays the start of the next test, although the two are
independent.

This patch does the GdbLaunch.terminate() on a separate thread to allow
the next test to start.

I've seen the tests run 34% faster with this change.

Change-Id: If4eb11e9d3e56dd06e5f7ca007f08d215e432725
Signed-off-by: Marc Khouzam <marc.khouzam@ericsson.com>