Bug 463976 - Implement a Console View based Terminal
Adds a terminal console which can be opened by selecting Open Console >
Terminal Console and selecting a remote connection which supports the
command shell service.  If a console is already open for that remote
connection it will be shown and if one doesn't exist then a new one will
be created.  Buttons for connecting, disconnecting, scroll lock, and
closing the console are also available.

Also adds a local command shell with PTY support. Also added some
syncronization around the connections in the connection type.

Moved the Open Terminal to console plug-in and renamed it Open
Command Shell.

Also fixed missing serial port feature from the remote repo.

Change-Id: I84282b04813471bb01385a328ebc144bdf0e6a28
Signed-off-by: Doug Schaefer <dschaefer@qnx.com>