Bug 464636 - Modernize TestsRunner Run launch
The TestsRunner Run launch automatically inherits the Run launch from
CDT but specifies the tabs itself.  Now that we modernized the CDT Run
launch and use CMainTab2, we should do the same for TestsRunner.

Also, remove the source tab for the TestsRunner Run launch; that tab is
only for debugging.

Change-Id: I4d412787f4e937ae776f01470f4c63746589ef75
Signed-off-by: Marc Khouzam <marc.khouzam@ericsson.com>