Update patch set 1
Patch Set 1: Code-Review-2

This patch should not be submitted to the branch yet.  This is a preview to garner comments from Marc K.  We need to have an official repo to get plugins from.  This patch uses a temporary off-site repo so build will work.  There are CQs in progress that are needed for an official repo of the docker-client dependencies.

Patch-set: 1
Label: Code-Review=-2