Update patch set 4
Patch Set 4:

(1 comment)

Thanks for the quick review.

 > * The current default is set to terminate GDB when then last
 > process exits. This setting will prevent users from seeing this new
 > feature.
 > Is it possible to show the exit code when GDB is set to "terminate
 > after the last running process" ?

For now, this feature was to address the multi-process case.  But you are right that it could be nice for the single process case.  We could look into an enhancement to show the exited process node even after GDB dies. I suggest we do that separately.  Note that the exit code is still visible in the title of the console in that case (as long as the process exited on its own).

 
 > * I like this additional way to restart the terminated process from
 > the same session :-)

Me too :)  But of course, that can only work if GDB is kept alive.

 > * Using the setting to "Automatically debug forked process" and
 > getting a small program which sequentially creates child forked
 > processes, don't preserve the exit code for all children but one.
 > This may be related to the collision of re-used group ids.

Thanks for doing that test.  I see the same thing now.  The reason is that GDB deletes old inferiors in that case; I'm not sure why.  I'll look more into it.

Patch-set: 4