Update patch set 4
Patch Set 4: Code-Review-1

I've investigated things in GDB and saw that when using fork() GDB tries to cleanup old inferiors.  I wasn't aware of that and this patch had made the assumption that this didn't happen.

I will post an updated version.

Thanks for trying this case!

Patch-set: 4
Label: Code-Review=-1