Update patch set 2
Patch Set 2: Verified-1

I don't know if it is a coincidence but both builds for this change timeout running org.eclipse.cdt.managedbuilder.tests.suite.AllManagedBuildTests. We should see if this change is the cause of it.  I'll run a mvn build on my own machine to check if I can reproduce it.

Patch-set: 2