launch bar: corrected edit dialog handling of adding of new LC
- launch bar launch config was not handling add of another launch config
correctly - it was just resetting itself to handle new LC. This cannot
really be done because this dialog is bound to launch group and new lc
may not be part of it and may not even support the mode current dialog
is in. Also that make missyncronization between LC stored in parent and
LC stored in getTabViewer which can cause other nasty problems.
So let parent handle adding (which only handles locatio change) and for
duplication we just open a new dialog and close current one.

Change-Id: I2dde5c3af26d901d1f4b18cad84a83a9857a6ca9