Update patch set 1
Patch Set 1:

Guy, I suggest that we remove the toOSString for now and commit the fix for the invalid condition. We can open a separate bug for the path issue and fix it in a separate patch once we know a bit more about that problem. I think the issues are separate, they just happen to be close to each other in the code.

Patch-set: 1
Label: Verified=0