Update patch set 1
Patch Set 1:

I tried the patch and Expand All only expands up to 5 levels. I did not check if nodes are lazy initialized or it limits it due to performance concerns. It takes a minute or 2 to expand the 5 levels on my project. Perhaps the performance concern is the reason why "Collapse All" icon is available via org.eclipse.ui.ISharedImages.IMG_ELCL_COLLAPSEALL but "Expand All" is not?

Patch-set: 1