Bug 466786 added images amongst other things
Added images including the error overlay for when connections are
closed. Made the resulting label provider public so it can be reused
by the launch bar.

Fixed a bug with the JschConnection factory which was creating new
connections everytime asked resulting in great weirdness. Added getting
list of services to the IRemoteConnectionType API and separated
out the three types of services.

Added open and close connection commands for the Connections view.

Changed connection type capabilities to booleans and removed unused
ones.

Removed launch config service. Found a better way to pass targets
to launch configs that will actually work.

Change-Id: I99d85f72d496f42d6ba790bd1b91943ea869c12b
Signed-off-by: Doug Schaefer <dschaefer@qnx.com>