Bug 467153 - Step one of making configs per target.
WTP needs a new config for every target. That breaks our architecture
of one config per target type. But that's actually already broken
since we can't really get the proper target to the launch config.
Depending on active, we always hoped the target didn't change, but
it can. So one config per target is correct anyway.

Change-Id: I14ba8413b9494a13f3496eed5535debbc330a56c