Update patch set 3
Patch Set 3:

If that restriction is to remain, we have to try inserting the breakpoint with GDB and then used the resolved location to determine if we already have that breakpoint. I am unsure where best to handle that error.

However, that restriction is not currently enforced when doing Ctrl-Double-Click on the ruler. e.g. Put a breakpoint normally on Line 15, and then Ctrl-Double-Click line 16, but type 15 in the "Line number" field. You end up with two breakpoints at the same location.

Patch-set: 3