Update patch set 4
Patch Set 4:

I've been thinking about another point.  The patch currently allows to edit the file name in two cases:

1- when creating a new breakpoint from the bp view (new action)
2- when pressing Ctrl and double-click in the editor margin

while it does not allow to edit the file name when opening the properties of an existing bp.

My first thought was that maybe we should allow editing the file name all the time.  However, after trying it a bit, I noticed that the marker does not always get re-positioned properly when changing the file name. Also, I'm not sure current breakpoint service would support moving a bp in GDB. I'm sure we could fix it but it is beyond the scope of this patch.

Instead, how about we don't allow to modify the file name in #2 (double-click)?  That keeps things aligned with previous behaviour, and only allows to edit the filename for the new action, which is the one that needs it?

All we need to do is check if the filename is null or not to know if it should be editable.

What do you think?

Patch-set: 4