Final LaunchBar 1.0 API. Lots of clean up and simplification.
launchConfigAdded becomes pretty important. Requires providers
to be able to find descriptor and target from the incoming
configuration if they own it. Also added handing for a default
when the target is null.

removed ownsConfiguration in providers since the launchConfigAdd
does that by returning false.

Added enablement for config providers, but that will need work
and some default property testers for it to be useful.

For the correct way to add in a config provider, check out the
Arduino change for this API at:

https://git.eclipse.org/r/#/c/49029/

Change-Id: I46b194a933c846d7d3e4eea8a0213c4ec324816f