Update patch set 3
Patch Set 3:

Thanks!

I was going to write a specific value handler for the pthread option, but this generic version is much better. However, this seems to be code from your company that you haven't pushed to CDT, so will it be ok from a legal point of view if I use it in my patch?

Patch-set: 3