Update patch set 2
Patch Set 2: Code-Review+1

We used bug 467109 to track these missing dependencies in the standalone script.  For a reason I didn't understand, it didn't impact the RCP version.  Marc-Andre explained that Tycho added these things automatically for the RCP.  I suggest we put a comment in bug 467109 to document this and explain why it wasn't a problem, but that it is better to add these dependencies anyway.

We should then update the commit message to mention the bug.

Patch-set: 2
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1