Update patch set 3
Patch Set 3:

If you go to the build in question, on the left hand column there is a "Retrigger" link.  This allows to start a new build from the same Gerrit submission.

I've triggered a new one now.  We have some intermittent tests failure which explains what you are seeing.

Patch-set: 3