Update patch set 2
Patch Set 2:

Each test version inherits from the base RunGDBScriptTest.  Our other tests are setup as a hierarchy of versions, 7.10->7.9->7.8->7.7 etc.  The idea is that let's say we need to tweak version 7.11 because GDB behaves differently, then we don't need to keep tweaking each version after that.  Did you mean to do it differently, or maybe it is a side-effect of your script that generated them?

Patch-set: 2
Label: Verified=0