Update patch set 4
Patch Set 4: Code-Review+2 Verified+1

>> 2- do we want to pre-fill the search box with the current selection
 >> when right-clicking?
 > Do you mean the start address ?  I guess this can be done as a
 > separate change not to mix the scope.
 > Note that once the Find / Replace dialog has been used, it
 > remembers the previous selections until Eclipse is restarted.
 
I was thinking of filling the Find tab with the current selection, as if the user saw a pattern in the memory, selected it and wanted to find another.  But that may not be a very good use-case.  Also, as you point out, since the dialog remembers the search after the first time, then it is better to leave it as it is.

For me, this is good to commit.

Patch-set: 4
Label: Verified=+1