MIRegistersTest fixes
- Fix getModelDataForRegisterDataValueInDifferentNumberFormats failure

Nothing seems to set the register we use to do the test to a particular
value.  The value happened to be 0.  This broke the octal test, because
it assumes that the result starts with a 0, but has characters after
that 0.  This is not true when the result is 0.  The fix is to set the
register to a know value beforehand (0x1234).

- Fix compareRegisterForMultipleExecutionContexts for slow targets by
adding missing massageTimeout.

Change-Id: I681acdeb08f8e3c2fdaf8dc9cf60801584da4423
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>