Update patch set 2
Patch Set 2:

(1 comment)

Regarding the renaming of getContainerContext(), 
I am fine to keep the name, not perfect but consistent with the name of other single process methods. the javadoc of the method also clarifies that this is used for single process, and on top of that the method will throw an assertion error if called from a multiprocess context.

Patch-set: 2
Label: Verified=0