Update patch set 1
Patch Set 1:

> > > Can you give the steps you imagine for this file once 8.8 is
 > out?
 > >
 > > cdt/releases/8.7 -> cdt/releases/8.8
 > 
 > Normally, since we only work on one release at a time, this sounds
 > good to me.
 > 
 > This year though, since we'll be working on 8.8.1 at the same time
 > as 9.0, we probably should create a cdt-baseline-8.8.target once
 > 8.8 is released.
 > 
 > And that makes me think we should change the file you propose from
 > "cdt-baseline.target" to "cdt-baseline-for-master.target" or
 > something.
 > 
 > What do you think?

There is only one baseline for a given commit in the repo. Since the baseline is in the repo too, it will contain the right content.

Even if that wasn't the case, in both branches it makes sense to use the 8.8 baseline (for 8.8.1 and 9.0).

Patch-set: 1
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0