Update patch set 6
Patch Set 6:

> Sure, why not? :) I'm in the middle of writing code. QtBuildConfig
 > is just started. No details there yet. Working on that next and
 > about to put a comment on the bug.

I don't think we should be pushing half-commits to master.  If you are in the middle of writing code, why not keep it in Gerrit until it is ready?  There is value in keeping a commit self-contained and complete.  With Gerrit, we get the benefits of having the code public and backed up, so no more need to push to the git repo quickly.

Patch-set: 6
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0