Update patch set 6
Patch Set 6:

> > > Sure, why not? :) I'm in the middle of writing code.
 > > QtBuildConfig
 > > > is just started. No details there yet. Working on that next and
 > > > about to put a comment on the bug.
 > >
 > > I don't think we should be pushing half-commits to master.  If
 > you
 > > are in the middle of writing code, why not keep it in Gerrit
 > until
 > > it is ready?  There is value in keeping a commit self-contained
 > and
 > > complete.  With Gerrit, we get the benefits of having the code
 > > public and backed up, so no more need to push to the git repo
 > > quickly.
 > 
 > Well, this was to quickly get the copyrights on the files. It was a
 > bit of an emergency.

Also, this will be an iterative process with very tight iterations, especially as I get feedback on things. I only have so much time per week to spend on this so I'll need to be pretty fast and want to check things in between chunks of time.

Patch-set: 6