Update patch set 1
Patch Set 1:

Thomas, how does this solution behaves on Windows? Is it valid to send EOT (0x4)?

I'm asking because if fixing it works everywhere, even Windows then maybe it's not worth putting the fix in the native code. Also given that termios.c_cc[VEOF] is highly unlikely to change.

Patch-set: 1