Update patch set 1
Patch Set 1:

> Thomas, how does this solution behaves on Windows? Is it valid to
 > send EOT (0x4)?
 > 
 > I'm asking because if fixing it works everywhere, even Windows then
 > maybe it's not worth putting the fix in the native code. Also given
 > that termios.c_cc[VEOF] is highly unlikely to change.

Well, I've tried to figure that you just by running the example. I run into some issues. By default on windows PTY (or the PTY-Streams respectively) are not used. Reason for this is in the static code of PTY which has console mode "disabled" for Windows per se. We force it to use console mode with -Dorg.eclipse.cdt.core.winpty_console_mode=true. But if I do that my simple test program immediately crashes. I tried to change it to a program which just loops infinitely. That resulted in the same behavior, it immediately crashes. I don't know whether anybody has used the PTY in the affected mode on Windows recently, I doubt it.

I'll try to find some other Windows platform for testing.

Patch-set: 1
Label: Verified=0