Grouping patch with non-stop and gdb 7.9.
View model code is in good shape, except for the delta stuff.
Grouping service allows to group as expected (views must be refreshed
manually).

Ungroup does not work yet.
Other services no longer need to be modified as we are using the same
IDMContext for each debug view element.  This has the limitation that
all elements must have the same children, so we cannot show a process
with a subset of its threads.  But that may be ok, instead we could show
a label in the thread node about its parent process.

Change-Id: I0d3a6aa037d0a1fce3795dda3daba7d63f042bb8