Update patch set 2
Patch Set 2:

Yes, please still open a bug.  We use bugzilla for feature requests as well.  I find it easier to use for feature description that Gerrit.

I'm familiar with Process Record and Replay, but not with the hw reverse that was "integrated into the existing code base of gdb".  Can you provide more details to what you are referring too (in the bug)?

Thanks

Patch-set: 2
Label: Verified=0