Update patch set 7
Patch Set 7: Code-Review-2

(19 comments)

I'm almost finished with the first pass of the review and it is in good shape.  However, there is some conceptual cleanup that will need to be done.  The idea is that whatever is part of CDT outside of o.e.cdt.dsf.gdb/gdb.ui must not make assumptions that we are using GDB.  That means the documentation must be more generic for example.  Also, we shouldn't assume that every debugger integration will be able to support this feature since it is dependent on the backend which may not be GDB.  Therefore we probably need to move some stuff out of cdt.core and into dsf.gdb; it won't be as generic, but it will avoid presenting a feature that is not necessarily supported if DSF-GDB is not installed.

I'll will point those things out as comments in the review, but I wanted to give you a heads up.

Patch-set: 7
Label: Code-Review=-2