Update patch set 9
Patch Set 9: Code-Review-1

I have brought all the minor issues up to date in this code. However there are still some bigger issues that need addressing. In https://bugs.eclipse.org/bugs/show_bug.cgi?id=472765#c2 I have a TODO list of things that I need to do, but there is also the relocating some of this code out of the core into DSF-GDB specific location. 

I am not sure it is worth expending much effort on this version of the code, it may be worth me addressing the key item of the restructure of the code first.

Patch-set: 9
Label: Code-Review=-1