Bug 478938 - Automatically add source files to Qt Project File
Added basic content assist for built-in qmake variables since there are
so many of them.

Added a bunch of new classes dealing with basic parsing and modification
of Qt Project Files.  For now the parser is a simple regular expression
matcher that finds instances of variables.  The modification class
supports preservation of indentation and comments that are tied to
variables in the project file.

Change-Id: I0539458d5c1cf29e6c9c1246e4e717e7cbec1b84
Signed-off-by: Matthew Bastien <mbastien@blackberry.com>