Update patch set 2
Patch Set 2: Code-Review+2

(1 comment)

I think the benefits of a working console outweighs the problems we have with dependency on globals, as long as this is only on the branch for the service release and not on the master.

Thanks Marc!

Patch-set: 2
Reviewer: Gerrit User 661 <661@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+2