LaunchBar: use o.e.debug control to manage lc tabs
- current wizard page tries to replicate functionality
of lc editor in regards to tabs but it does not work
the same which creates issues for contributed tabs
- for example it never calls "activate" on the tab
- we also have rendering problem - pages are not resized correctly and
show up empty until resized
- I switched to use standard control called
LaunchConfigurationTabGroupViewer to show tabs. Hack with fake
LaunchConfigurationDialog has to remain to interact with this control.
- If we move it to o.e.debug we have to fix these classes to remove some
of the hacks and expose proper APIs to re-use this control

Change-Id: Ia39706fb5972883167f5f39751985d97e1815c86