QML Parser Tests (Work in Progress)
Thus far, added utility classes to handle assertions of the parse tree
that is generated by ANTLR.  The goal with this is to make understanding
the structure of the parse tree during a test easy.  This also
facilitates the simple addition of new tests without having to copy
common code.

Also, began fixing up the grammar to at least support the "HelloWorld"
base project that we are generating.

Change-Id: Ifc41bf128d14e4cfd043d1a11d2c848ffb24662f
Signed-off-by: Matthew Bastien <mbastien@blackberry.com>