Improve robustness of downloads and add license confirmation.
Bad things were happening when the downloads of the tools and sdk
failed. Added retries and changed the order of the downloads so that
partial downloads aren't registered as complete.

Also added license confirmation dialog to make sure the user agrees
to the Arduino licenses before installing the tools and sdk.

Change-Id: Ie8f4fcd041d8e89195bc7d3551c63fd3270881ef