Bug 480238 - Parse QML Object Literal and Members
Acorn-qml can now parse a large subset of QML syntax.  The only two
things left to do at this point are to allow QML Object Literals in
Property Bindings and to allow QML contextual keywords such as 'signal'
as property binding identifiers.  Both of these require lookahead which
acorn does not ship with at the moment, so this may be a bit of an
undertaking.  Also, added a whole bunch of tests to parse the new
syntax.

Change-Id: I0950fa29265c8337b5c9bfc0a1ec0c3ba8267426
Signed-off-by: Matthew Bastien <mbastien@blackberry.com>