Bug 336876: Grouping feature
Delta stuff mostly works but not perfectly.

Resume and suspend work on groups in non-stop mode.
Multi-resume/suspend need to be tested.

Grouping service allows to group and ungroup as expected (views must be
refreshed manually sometimes).

Some elements in the debug view can now repeat: a thread can be shown
more than once as it can belong to many groups.  We use the same
IDMContext for each repeating debug view element.  This allows us to
keep all other services the same as they can keep using the IDMContext
without any modification. However, this approach has the limitation that
all repeated elements must have the exact same children, so we cannot
show a process with a subset of its threads.  Instead, for threads that
are directly children of a group we could show a label in the thread
node about its real parent process; this patch does not do this and does
not show any info about the real parent process.

Change-Id: I0d3a6aa037d0a1fce3795dda3daba7d63f042bb8
Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>