Remove visualizer plugins from Stand-alone RCP debugger
The RCP stand-alone debugger does not have the multicore visualizer
plugins but does have the visualizer plugins.  That means a user can
open the Visualizer view but nothing will display.

We could either add the multicore visualizer plugins or remove the
visualizer plugins.  To follow what the stand-alone script does, this
patch removes the visualizer plugins.

Change-Id: I90075c7fc5c87f21c0b697f81c45abf1d0395218