Update patch set 1
Patch Set 1:

The Qt plug-ins will eventually be in the simrel repo and I expect people to be able to install from there. At that time all it's dependencies will be there too.

Adding these dependencies to the feature is a pain to maintain. In fact, I'm about to remove the ANTLR dependency all together. Would have been bad if we forgot and left it in the feature and everyone installing the Qt plug-ins would get the ANTLR plug-in even though it's not used.

But, yes, I'd rather see the category.xml/Tycho mirroring done for the CDT repo for these early days. It's what I do at QNX to mirror. Works well. We just never thought about it much for the CDT repo since we didn't have dependencies like that before.

Patch-set: 1