Update patch set 2
Patch Set 2: Code-Review+2

In the case of remote debugging, we don't have a console in Eclipse, therefore the exit code cannot be shown there so we don't bother setting the exit code.  So, the test should not be run in the remote case.  LaunchConfigurationAndRestart is already not run in the remote case, so we're all set.

Patch-set: 2
Label: Code-Review=+2