Update patch set 6
Patch Set 6:

(16 comments)

I took a quick first pass and put a few comments.  I will continue working on this as I find time.

My first concern is that using this new hardware reverse debugging requires a specific processor right?  If this feature is only available to a subset of people which have the right hardware, I don't think we should make the feature as accessible as it will cause confusion.

Is it possible to switch between software reverse debug and hardware reverse debug during the same debug session?  If it is not, then I'm not sure the old toggle button is the place to put it.  When must the user choose hw vs sw reverse debugging?  Can we first start the debug session and if hw reverse is available then show the option?  But maybe that will be confusing as well.  Should we use a preference?  Or maybe a debug view menu item?  or context-menu item as you have now, but not modify the toolbar button?

Although you have handle the toolbar button nicely, I still find it confusing for anyone that is not aware of this special hardware they need.

Patch-set: 6