Update patch set 6
Patch Set 6:

Hello,
      Thanx for the comments, Yes the feature requires a minimum processor version but 
will be available in all succeeding versions. At the moment GDB 7.10 already has support
these features. Yes it is possible to switch between software and hardware reverse methods
during a debug session.
The Hardware debugging has the capability to capture a lot of traces without a lot of overhead
so the user can afford to keep hardware debugging ON and when he runs into a problem, he can see the
traces to pinpoint where he should further investigate. Actually there is already a preference menu
contribution in this gerrit, where the user can select between the hardware debugging methods
"Branch Trace" and "Processor Trace". "Branch Trace" being an earlier version of reverse debugging
 methods would be more available for some time before "Processor Trace" takes its place. Currently
GDB also has support for choosing the best method based on the availability which is why there is
the "UseGdbTrace" option.
Actually the hardware reverse debugging has been integrated into the software reverse debugging
framework in GDB, which is why I put it there in the toolbar along with a preference page. Currently
there is no way to query GDB if the hardware reverse debugging is supported or not in the underlying
target.


> (16 comments)
 > 
 > I took a quick first pass and put a few comments.  I will continue
 > working on this as I find time.
 > 
 > My first concern is that using this new hardware reverse debugging
 > requires a specific processor right?  If this feature is only
 > available to a subset of people which have the right hardware, I
 > don't think we should make the feature as accessible as it will
 > cause confusion.
 > 
 > Is it possible to switch between software reverse debug and
 > hardware reverse debug during the same debug session?  If it is
 > not, then I'm not sure the old toggle button is the place to put
 > it.  When must the user choose hw vs sw reverse debugging?  Can we
 > first start the debug session and if hw reverse is available then
 > show the option?  But maybe that will be confusing as well.  Should
 > we use a preference?  Or maybe a debug view menu item?  or
 > context-menu item as you have now, but not modify the toolbar
 > button?
 > 
 > Although you have handle the toolbar button nicely, I still find it
 > confusing for anyone that is not aware of this special hardware
 > they need.

Patch-set: 6
Label: Verified=0