[launchbar] Support launching without a target
- we should not make assumption about the target and let
delegate deal with it. For example if no targets are
created and you run for the first time, delegate can offer
to create a target
- also if launch bar participants don't care about targets at all we
should be able to still show modes, etc
- so I removed filters on null target and let participant deal with it
(note null target is not the same as Local). To avoid NPEs null target
is actually an object ILaunchTarget.NULL_TARGET

Change-Id: Ie2a4d3430284adbb137a4565d9976a0ba554d41f