Bug 421166 - Provide DSF/GDB debugger support independent of language
This change relates to the issue above. It's a fix to a bug when
debugging with non CDT languages: When the debugger tries to open an
editor for a source file with an absolute path, using
AbsolutePathSourceContainer, it will call
SourceUtils.findSourceElements, which returns a one element array with
null inside, even if the file exists. This is because it tries to create
an ITranslationUnit for a CProject, even though the CProject does not
exist. (The project exists, but it is not a CDT projecT)

Change-Id: I375e6a4c8677f86eaceb3d57656ea18192b0d5e8
Signed-off-by: Bruno Medeiros <bruno.do.medeiros@gmail.com>