Bug 482938 - Pass LaunchTarget to launch delegates.
Creates a new standard delegate that implements the targeted launch
delegate interface introduced earlier. Also provided a standard
launch object that stores the launch target to be launched on.

Adds storing of the config to target mapping in the launch target
manager. Sets that mapping as close to build/launch time as we can.

Change-Id: I78a1412af043f0be240d5617fcfc08b81c3fb3ea