Reply to a DSR(5) request for ready status.
When TM Terminal sees this:

   ESC [ 5 n

It replies indicating that the terminal is OK:

   ESC [ 0 n

This escape sequence is documented here:

   8.3.35 DSR - DEVICE STATUS REPORT.  Page 40.
   http://www.ecma-international.org/publications/files/ECMA-ST/Ecma-048.pdf

Motivation: I'd like to switch winpty from using DSR(6) to DSR(5) as its
ESC flushing command.  When winpty sees a bare ESC, it can either
synthesize a VK_ESCAPE keypress or wait to see if the ESC was part of some
other escape sequence.  It tries to deal with this by generating a DSR(6),
which should flush out the remaining bytes of the sequence, if there are
any.  DSR(6) is suboptimal, though, because the reply collides with the
sequence some terminals use for F3 with a modifier:

    ESC [ nn ; nn R

DSR(5) generates a DSR(0) reply, which does not have the same problem.

DSR(5) is supported by all the terminal emulators I tested, including
xterm, rxvt, rxvt-unicode, gnome-terminal, konsole, putty, mintty,
OS X's Terminal.app, and IntelliJ/jediterm.

Signed-off-by: Ryan Prichard <ryan.prichard@gmail.com>